Vars 'w','h','d','currentpicture','currentpen' cannot be shadowed.
As these variables now are prioritary and are not stored anymore in
the metafont structure and they are bound to execution context, being
destroyed after running WeaveFont code, declaring a variable with same
name would cause confusion: the declared variable would be completely
ignored. Therefore, it is better to transform this in an error to
prevent confusion caused by values stored in these variables
disappearing between each charactere rendering.